GIT/rirekisho-parser/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
openai>=1.14.0
python-dotenv>=1.0.1
pdf2image>=1.17.0
Pillow>=10.2.0
